fix: refactor: moved the xml-parsing functions into separate files under each profile
