+++ src/mosModel/profile1/mosObj.ts
@@ -1,7 +1,7 @@
 import { MosMessage } from '../MosMessage'
 import * as XMLBuilder from 'xmlbuilder'
 import { IMOSObject } from '../../api'
-import { Parser } from '../Parser'
+import { XMLMosObject } from './xmlConversion'
 
 export class MosObj extends MosMessage {
 
@@ -16,6 +16,9 @@ export class MosObj extends MosMessage {
 
   /** */
 	get messageXMLBlocks (): XMLBuilder.XMLElement {
-		return Parser.mosObj2xml(this.obj)
+		let xmlMosObj = XMLBuilder.create('mosObj')
+		XMLMosObject.toXML(xmlMosObj, this.obj)
+
+		return xmlMosObj
 	}
 }
