+++ src/mosModel/profile2/ROList.ts
@@ -4,8 +4,8 @@ import {
 	IMOSRunningOrder,
 	IMOSROStory
 } from '../../api'
-import { Parser } from '../Parser'
 import { addTextElement } from '../../utils/Utils'
+import { XMLROStory } from './xmlConversion'
 
 export class ROList extends MosMessage {
 
@@ -25,8 +25,7 @@ export class ROList extends MosMessage {
 		addTextElement(root, 'roSlug', this.RO.Slug)
 
 		this.RO.Stories.forEach((story: IMOSROStory) => {
-			let xmlStory = Parser.story2xml(story)
-			root.importDocument(xmlStory)
+			XMLROStory.toXML(root, story)
 		})
 
 		return root
